Fix '.' imports
