+++ dev-packages/private-re-exports/src/package-re-exports.ts
@@ -17,8 +17,7 @@
 import cp = require('child_process');
 import fs = require('fs');
 import path = require('path');
-import { parseModule } from '.';
-import { PackageJson, ReExportJson } from './utility';
+import { PackageJson, parseModule, ReExportJson } from './utility';
 
 export async function readJson<T = unknown>(jsonPath: string): Promise<T> {
     return JSON.parse(await fs.promises.readFile(jsonPath, 'utf8')) as T;
