+++ packages/filesystem/src/browser/breadcrumbs/filepath-breadcrumbs-container.ts
@@ -16,8 +16,7 @@
 
 import { Container, interfaces, injectable, inject } from '@theia/core/shared/inversify';
 import { TreeProps, ContextMenuRenderer, TreeNode, OpenerService, open, NodeProps, defaultTreeProps } from '@theia/core/lib/browser';
-import { createFileTreeContainer, FileTreeWidget } from '../';
-import { FileTreeModel, FileStatNode } from '../file-tree';
+import { FileTreeModel, FileStatNode, createFileTreeContainer, FileTreeWidget } from '../file-tree';
 
 const BREADCRUMBS_FILETREE_CLASS = 'theia-FilepathBreadcrumbFileTree';
