+++ packages/scm-extra/src/browser/history/scm-history-provider.ts
@@ -1,5 +1,5 @@
 // *****************************************************************************
-// Copyright (C) 2018 Ericsson and others.
+// Copyright (C) 2020 Arm and others.
 //
 // This program and the accompanying materials are made available under the
 // terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,20 +14,14 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 // *****************************************************************************
 
-import * as prom from 'prom-client';
-import { injectable } from '@theia/core/shared/inversify';
-import { MetricsContribution } from './metrics-contribution';
+import { ScmProvider } from '@theia/scm/lib/browser/scm-provider';
+import { ScmHistorySupport } from './scm-history-widget';
 
-@injectable()
-export class NodeMetricsContribution implements MetricsContribution {
-    getMetrics(): string {
-        return prom.register.metrics().toString();
-    }
-
-    startCollecting(): void {
-        const collectDefaultMetrics = prom.collectDefaultMetrics;
-
-        // Probe every 5th second.
-        collectDefaultMetrics({ timeout: 5000 });
+export interface ScmHistoryProvider extends ScmProvider {
+    historySupport?: ScmHistorySupport;
+}
+export namespace ScmHistoryProvider {
+    export function is(scmProvider: ScmProvider | undefined): scmProvider is ScmHistoryProvider {
+        return !!scmProvider && 'historySupport' in scmProvider;
     }
 }
